==> Semana-6/3-model-test.ipynb <==
import base64
import cv2
from PIL import Image
from io import BytesIO
import matplotlib.pyplot as plt
import numpy as np
from keras.models import load_model

b64 = "/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/2wBDAQkJCQwLDBgNDRgyIRwhMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjIyMjL/wAARCACgAUADASIAAhEBAxEB/8QAHwAAAQUBAQEBAQEAAAAAAAAAAAECAwQFBgcICQoL/8QAtRAAAgEDAwIEAwUFBAQAAAF9AQIDAAQRBRIhMUEGE1FhByJxFDKBkaEII0KxwRVS0fAkM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwC4VpNtWSlNKV6pwFfZSbKsbKTZQBW2UmyrOyk2UAVilJsqzspuygCvspNlWNlJsoAr7KTZVny6TZQBW2UbKs+XR5dMCrso2Va8ujy6AKuyjZVvy6PLoAq7KNlWSgGM456Uvl+1AFbZS7KmYKhG44z0yP0+tRmVFieVgQik5JB6D+L6e4pXSATZVf7RuuDCiMWVgH46Z5B+mM/p1qe7lSKxeUOFJUhMjkt2GD39qy7Szvb4215JPI8LKP8AVTBTg4zwB0z15BGOOgzE5O6SKSVrs1JD5UZfaWx2XqaiuJmjWIwhXDMueeqkgZHr1FOv5GguLZQ0KxtuMnmA/dGASPoCe/TnoDirqcrybYgFVTuRoWdlkkPGO33Tgg4znPByKzrVuS5UIXNIJWRcamyak1rEYnfKBASRlt2CvUcnIHoPwIqxHrUSkRymPzWBVQ2U+ZTg5HOOnQ8e9ZEtt52oO3ngO8siLtiAkyyttGemCfXHAxxXHi8VGS5Ybm1Kk07sr6nLbvcSRTWsLzDmMliS2QQfm5BIAyDjg454qvJ/ZsKSSx2Bjix5MrPIJNzDPAzkE7l5xwNo6ZOYYHY3aOEjlPzApJCyvEQx+8PubuR1yeB7VHfG/bU2tywSyRwzPKepCrkgd+WJA9COea893b3OiOxYSdpb63ETqOflKjKs2cg5JDMQVUcZHABK8Ulqbqea+k022WRFCN8ibUcEkYCk98npwevqRmzPHaXaXMTMDZx+YiOc7znZ8qk8YYDpk5z020+PXtTitprlf3UICrv2nBIGVUccHr+XtVcja90V+5bvtGhv5J7kXkwkSQiRYoQyE5OAGG0AHacHkd6sXmjxWccUlw0kbO3nNHPbI4LBCgGWXcEyCOcquRxnJGfp1/Pbw2d299BbBZQYopnIwMjEhGxsqCBznd/s1sazrs99pVppz3UZtixM13cK/lFx5jbo12s4z5hG7G44Q4XmqXMmhWTRialFbxI8ltEXZ4xLMVmDgks2M7WZuQdxDEsCfQCs9NLt3dZHhvLpmG1UL/dAG4kd8c5HUc+1akCifTY5bi52GW4XfcPJ5nOGwX25P3jnnnnODkAtu7maGJ3tFLwWxw91FGdrZOAWcgEAkDaDjv3zk5nctrQ9P8H2emtof/HpHYsuWkWUiMBRwZOMbh2LHupHarvhLT7S30W71VrBJnuJ5ZhGjrv3BiCgYkZwwbklRk9q8x0O21S7tri+gvcRh0S53u3nNHuQbQcfdO/pkBgp4O2vVAscWi29paXKQXElspSTC/u2wCWKsMNyefr1yc1nUqTejkVGCSukct4phW9vF1ZtJnskLeWkd9OHEpQE5KAttG4Ff4lPDcbiTTt7DFhc2s9+bcpaFYGTGN7DJQnk4LbgSO3cbyCXvhSecxSar4jnvpGDebbxRkIpYHcFdmIVd2OiDIGBjtsWNhNb6WLa1dRaRSbgzuW+c53H2zk8Lgcnjk1m7LZgtTY8uk2VdMVNMVfTXPJKWyk8urpipPJouBS8uk8urvlUnlU7gUjHSGOrvlUnlUAUvLpPLq75XtSeUfSgCl5dHl1d8r2pPK9qYFPy/ak8urewb9uDnGehx+dMDKZli2sGZS4yOwIH9aLgQeXR5dSxs2FEsZ8wn5ggJCZPGT+Wfz6U6ZlhKqRl2DEIPvNgc49T04ougIfLo8uqk98skvkpPHE+8AISCxwQGHB68ggdTgjBBqwf30QhVZQyMIyzAllbA64BxwTk5HscEGpUkx2KrtK88f8AopaMOBncQF5IJIxzyBjj3OKsvCZJFBQBVO7LAHJ9vSo7E2dt9s2t+8DSzSgFgDhm9e/GCQMcfSkvtRC2Uv2KKVbneUiSXH7wlmAIzz/CTgjg4FZyqwjo2UoN7FC90+4nvJnjKlJB5WS/yrwOvoQc4wDyeg6l/kzJP5Mt5lBtG44JZzwV2k5+6MjrkkntzpQy2U+pSpG6yLM4EeYypDqMkgj7wITOeANqjnNc1cQzJqFu1xAwe7kk3gPHmWP5PuN0+63XgEZ6DNc0sRBu6NVTfUs3FxM1m1neqHaXJy6qpUdVyDkHtkDkDJ6jiWWWPT4RcRTQSQWq+WpmmLO2RuJz/CPkYZ55GABgism4hupplaK7hDw7JHuTI4YISCv3hwORj6DHYVakuLBleS4njuZEtjKbaOJY05lKnIwMDLqw5JILZHzCsHiZ62NFSiOtLy21i7tUvmk2binmQ53fMcbSoBBBIVu549azbtZYNUTyGMsE0YfzZCeD90qp75KtwRwAvzdRSwSW5jmktnuPLVo97RPtMWXOMnZkn5cAjAB5waq6le2dmLA3q3txb/ZY9yqNsczMiyMNwAKkBxx1wVzxXNKU576mqiokVxK9zPIbdA9qsDeWWBIMgckrhtxBP95TyPQEgrb2N5PZm8tYvJKny3MQLbVUcsQB2LcgZHTHc0+PwzrGrHNp9vsZin762aAgFA5woLsN2FK8MecEHpmrEXhf7JYy2lzcW4jPmSPOC3mokZ2yKGMY3Hf5R7Z2kDh2NTZdyjlr23cTz/ZriFkR3eMFfIDR7gofkjbnzDgDAHzYxjFdDpkNxrMYg8xru+OGt0W6iLOcZJ8tiGIxznqRk4Hff0y/8PWGrxLHpkqTj/lvlCmMEFh8zMqHLPy33QhGQeOr0O0GnwWqQXFx/qBvzIxXDEvt2A7S2TjcRkDqaic11RUYmTofhu1iuLi/3zLbNAiyXbfu0K4y3lvhdy/Kvz84wSCAaxdB8N6hfaXJdvc2RjlVjdF58O8pK7xIzRsFUKM45JzknDYHpWp3d1FFIWiaP92WUMSeg4z9TXj3iyGObUrS3j+2wwRWeSjvgJM5Mm9RvJBy2CrKOF4x8pGcZt9bFSikrlS/0+bS2vJC6R2FvjcZofNSQh9gCN5aDBIONnPHoCRFemXzbu3i0tInjcWlz5ONsZLEHcdpwcqQCpH8XJzTo5oIo7iOa0t3gutnnbQdw25xsZy7JnIyQecDtxVZDGk7W9mCIJZS0cMztIkbN8uAuDkYxyQTwPx0U0zK5c0qya8tI7O2YD7XqSAPEBM0SnCs3YrgKW7Hb/snJ6jw14CMlrJJ4hd7WzZ4pltRcK7Fow4O5gnQ7jgIejY6gGp/DMFxBEdT1G3jSaPmMR5QAiMRLhQeDsyNvY9uBVi+a/uI1Z2ZYi3Q5LEfTsPxqZTeyNbXsyxdrpdvENM0culkmWl83c3zBkYFGZvlHBzxg7vxqnpafaWaaR0gQ/3VwTxjO0dTgdT+OarM6W1szyLvllO1FBxuP19B61q6TZTRBPOiL3UhA2AYCDtknp9KzeiKSvoasHh6zvERZFl8hTu2eZgyn1bHQew4rVFvZWm3yIl3qPlUfdUe3+fzqtLceSn2W2Cs23LnpgetZs91LBCFb5mbLEDrjpk+gxWV3I15VE3jBUFyyWsXmyhhGCAzAZCD1PoPU9upwASOd1LWJZrWdbG9lPy70EQb5dv3wHYAtgMGb+78oGQSRBe6/a3vhq1NxArQBJVcTyszCWNPlDMu3ruUk9PmPTrX0jq2VzxVA6xEWSNZI2V0YAqynIIPQg0vkGuS0jxZYWa3MVs0t1ZRshBMjMYgygkKWUEqOVAIB+Ukn5uPQ0spJD8qfieB+dUqqtqTysxhFuLDaRtOOR14zxS+QfSqus6Nf2ut6fqkakRKxjuHWZQfKzkAjHIyW49FHfkRC/u7vW5LS2R3tpEaNHC4Eb7cglgO5GPx/ChVYvZhyssqYnmeFXUyp95O44B/qPzp7RhAS3AAJJPQAe9czdaRqul39i9hZ/aV84MTbE4B3fOhJzgfw5Pbr0rptdtZTBG/k+fbqSJLcHBlZsKgz0A3NnPbAPaqU3YVhFjDoHUhlYZBByCKZhTMYl+Z1xuA/hBzgn8ql0h9T1PTo3Nn5dzJIiBvLJjwyht/BOAAWA+bkrjgnbWhZ+GJY7m2mu7yV5Y8/aIoI2+c7WAXJc/Ku4clcnGepNRUxEIO0mXGnKWxlOY0jD53LvCfLzyW2/z61BeXVvZNGszYLn8h6/n/AJ4rr4vDelaPbpJHDK8iZZGmd3AY5z8nCDgnkD1PvWJdT2nnbbqzsxKTtWTygzYBzgHGV5rH66nsinRa3Mu6sxPPGS6qqo45VTyQB1PI79Ovr60Psd1/ZIInWK5RpAXeZlUDJH5YAxx7iti5tPPvhJCkShAwUyZz1BIHrnHrxgdaxdXg1Kae6thp9wYootskoUhSCpY4YgrjAHzZ4NV9YvqHsy7bW1knmXHlJAwdy2WwMA7d3pyBn2yfU5p3+qW8Oox2ZVvNZlw3lg7VDguTzkDA9B0znGCOftppLuxug6SB408s/wCk5aXhWVmJOAMFiCMrgnHQZrRrqMKW19FEkaumJJZEZlILrsxkZJyykHI6Hng03XlsCprc6bV49JsJbR5IoENxKIizx5UJgA8HAAwqj2BOKz9T1421zdSRQQmaHiMyN0AUnpkDPX3G4gZ5xFc6Pq09gtwt1O1w0ZWNBkseQoGcHHB68ep4yRVktNT8PQQy3l9axy+U0pW5cPJEflAIUkbidrDoQMctg5rGVeb1RoqaLAvor240+6m3RTXX7uIRTCPkEA5bbu2sc9CceuTxBPHc3lkt3eytCGki/wBagYEPk5VR6DaTnqQTkFa1NCs9FtdPs5r+LzdRS4czILhXWORW45QHaVQdCcHOD2xX1hPDwswltLqkscCeWohdAVYN1GFxu3ck9T36Vk3dFpGRod0mk6hdzTlpLeNWDTxyEckcAAfIxJOMHvz0UmrFy9tZaNPaWSzBUnMTs42ywwuiDcyY3bSxU4A5C84NS6FaxXTbbywuprX7TLEJImkZoi8a/eMTLtQ4HO1gcnITvtaHaXUmrX95e2oeCVobmOSQ4AlEQ+4o4IAfGcdVH3iMripWVmaOOuhzi+Gb5bqNrfTprcSeYZLbz8tsRlVthYjqXB5OCFJGMqDtaT4QSe1luNUl+yG5BEtrCdqq5JGckcnJyMcckDKnB1Nci1eLVNOk0axQeXaSxRkyouxjsYKEYAEERYGD/Fk7QuSlvcaytq41CaO4uIpxGblDtEyMwHmYA+XAJBHTK9ec1DldXGl0OBj03UgbFL8zsscieYfssjLEHAfBXjevRGwQQxK44XPW+HtOh8h0ubO3UiWWJrdIAEfIRMrnkDCHrnIfJzxWw+n3c1uzbCIpLmOQbnxlQU3Dvg4DEDpnHqTV6PQbhjLJGHdC27yhHg42gYzk56frWcqi6GiiVlMzatJGju37iEbtxLMQX9P/ANXWuN8VytqGrRaXGnmzPvhZZoAIIwsW9pTK2FDKHOFU4GMkjgHvhP8A2ZeXM624BaCOJEeQZVwXJJHUDDLjjtUmjWy3FjdC7UtDM7SOi7QvKgbMY5XAxyPXk81jz63ZfKeCz5l1Br3T/ItAreaAlwAIFXHzHBJAywweSSQoy3Few+H9QYWa3UcCh0DRlGG4MVYrnt1C5xUlxoljJpN5pN3eC6+3SJ5knkhmCqkaHaCSNx8vO7qMg9RmtJbC3so1jsoVjt2Lt5ZJYklixPOe5PenOopJIFFo8z13xvqd7JLbzM0DpO4kVHzxlMKfXBQ899xHA4rMt9H17xNdSX8cZaGRvnuZm2qDj25x2GBjt2r0C70uwleUmztMySb5f3Ybc3PJ4xnk88nk1pG3A0+NVAjt41CiNerY6DA4H09KSaWxPI3uec2vga/mvVinnilgJxutmJP/AI8ox9a6e0stO0yNoLRI1kDBS6HIzkkcnJYjnk9O1Xb3ULiO3+ywBY2k4YR/eC+hJ6Vn2Vs1lGJpVCL/ALXP5fpTuxqKRoK0NnAombMpG5YVAG1fUntms641Wa8ZWkBa0BK7FO0MB2A9Peo5tlwsktzIzGTiNFHJPqf8KtW6RRW0KtEJJUGFiUdO/J+vWjRFWuaen2WZRfXiqsarlc8ADsqj+taDXSXcjrGyKxBPIACj6VmpZXF3891NmXqsQ/gFJb2ZW6ZbRhNLggPuyFqHZlrQuf6HbWbzM7sinLuOhPpzWLbvcXMRdLV1Er5Mk52qF9euT/nucVcl0mcxRpMr3AhbccYWMHOckkf0NQzyXNxKA4jAA2pGvQAdPpQgep0niLw/9pu/3unpNIkQRGiZ1GwHI4BAOCT9Pxpmjabe2FtLH9itJElm+0NEypEFYBQu0DjgoD068+9ctq/iVZYHWw1S6a5I3JL9nRkz6EEA9PT179K5cXOrSuJLnUnkbOThEAx6YIP6V6Drcmjl+pxKg5dD1aPSYLmTzryzstyx+UpuSmxEyxCBQAGA34GQcYHfJNh/FWl2CsJbkyuQd5X+M/z/ACrydpJHb5p5G9MscD/ChAqjP3q551k2bxw76nc3/j4TRLDZWgjU/elZQWP51mjxJdPk7UZm4JKr+HQVgR5fkKpA7AVdRwABjJ+mP/11lKvJbGqw0eppJq14ZBIJVBHQYP8AOtO2129CMplDAsrc5JGDnAOc4PQj09Kx4grbhuO7r26VMkTRnbFkHHTHWs3iavSRX1an2OwtfE9rbmFbTSkt4YskLGw+TJ5IXgZ/zmtiXxFYTWgbdcSluWhfgrj/AHRgn2z+VefrI/Hm4DY4xz+tToVVt8cpB+v9Ky+sVU73uV9Xgy7qmt3F66qAyRxHciuu3d9ayrSeTUZjDMWil6lHxh/TB7fhU/2u4W5KzEeUeFGMfrU/nER7kIOB0JraONcd4mUsInsy/p9ndvGsW/8AfAkFj8uPY+v60a5qFx4f06QqrhZThvmyu7H6CsuG7/fESuwZuvOBVtwgjZFXKt95T0YetP68ub3o/iR9UdtGeX6Zr+uPfC5bS7m9sJXAci0MoMYLAhT04DOOvseK9EOnWq+VJYxwwI7CWVBCI2lGDgMOCOTnp6is6+iuEIa3lkK91Pb6U6A6lNsTyncMpZdwwCBnOCfofx4rb67Kb0iL6oorVm6IpkxM83lDsY/LDRjjoxGQOMnB5/AVnSz6dLHcXl7eF5I0dnd3DkL1bBXPp0HpT4rK2jSQ6hZWs2VXajSSkg9xgMPUflxVe78PaVcRSxfZxDHIGXbE2NoPoTn9c1tzyZlZIsP4Yt5S0rOgiLZYoN2/gDk+uABnmsCaz0KHXlt7cXF48DgPEzr5W7+62RyQccYx2PcV1RS2+zC2tZry1DHDP5qyEKRzj5Rg++e1Z9r4O0mCVGsb3D5xm6+VeemNoP6kVEqlRR2KjGLe5o6bBNF9sLKPLnuTMqq+MAqud3A53BvwxV+302GOyLW6sgZgSCxOSFC8c46KPyqs2m6xHCdsK3ECuUDW22UHr6ZPT1A7d6b/AGg8WLadZImj4MbLtx+Haub20krNG3sXJ3RJceazlCCXY/xHcxpX09obYXV1KFRcnbgZPbP+H/16QX0RBy7EseSfT8Kiea2EyOELqg+VSOM/pmpdRMpUpdi29xcztB5T+RAnIkK8yH0GRwO/v7c1UfXprT/RoYoSoO4zSRDeM+nbP1z+NV5b53k3iFwwGBk5x+nH4VUZmwTsLOfXoPX61PMmaKk0KNRwjEKzkHO5zwxP60jTSSx+Wgd1HL4HAP8AWkjtjJIDMB67VNbMVuCiqNx44B6UuZD5GZvnGAfugwGAMnkt/hT42uDGU+VWblyT90e57/54q5eQthflXIP3iOP/AK9VPNw5JYu2MbjiqTRLT2Kk2YdxJKop6leX+npUMMlyylUV9rjbGuOnqa0o4YbkgzSNuB4B+6B/n61fhtLclXWZVYdcPk4H4cflQ6iQezkzKWyGlWqNMI3mfrgAbfx6mqdxC+pTiPYVjHBxwPYD1rp54bJhiNMg85MY5Prk1PC1tbxJtSMuFG7PT9MVPtB+zOQutCfT4DLNIuWxlF+83tntUVsLqZHNvEqDP3cZOPc9K6i8e2nMeY4VVTnCIcj9MUyADJdX2xr0TGB+NPn01DkMuDSZxCv2otKp6xjuPf1ratbO3htmnuWCBP8AVwgYHHc1Xe/nfcodlHrGSP1rKlnCJjazS+shJ/HrRdsOWxcmuN0cirI7hyTgL396xpIm3kAyE92ZQOfp+frWtYJfOBGGVT1xt3MB7noBTrmyKgSXDeaVOMIc8+9JSsPlueeeSw+8OPypwiyVKn8RWiYvMXOxcjsOD+VRiJSc4GPoavmHylZIg+MKSKk8lOvIx6mrIg2klTxjAGcGhIzIMv1B7c4pXHYACwWMhQQODgCrEMJJVV4B4JH/ANc00oOCckZ65p0e7OVxgdQTiobKRKI3j5GBwQVP+FWIbkIvlsMMRjOaLazuZwpKOsbZIc/KpxnucDt61MlriPPnxiQHAQZJx65UEfrTVOUtkJ1Ix3YrurR7lGWHJxSJNyMdc9fWrSadxtHnPGB8u4BCP/Qqsx2C5RvKiQqMZC7t3GMkNkfpWiwk3voZvF01tqUlJmdUALMWwoAyc9MVMUdFbBAK/eRztbPsD1/Cp3NpbAxzSqAWyY8nrjrtHTgVG+qxIALe1Occ7sACtFgo7yZjLFv7KI47RnkXqVK8kL+hBxV4J5KIZpFUYO5WfhvoBg8fjWQ+qTXORE7Pz/yxIVFOO7E/1pqRTSsTLIV5yQnJOf8AaP8AQD61oqFJdLmUq1SXWxqG8trdoyhZ2GQrEhMg9enJ9OlQi7uJOEAhjByMDH6Dn8z+FVFUL8sKlWbALMOT+J6/nThIoGHYHnAyc/y5rZRtotDJu+5bjcICoY8fxEcn/PtSCRyccD6rzVcTqNqrjcR6E5FQSXgMyLnJ9A3T8O1UlYLl5mO4ZkJB6CnLMykYYqAOQw5rPLkHJiYenGP506PczswDLjnIHf04osBrRSzROJIJWjkHIZMjr7cevbFayeKHKGPWLOC6iDZBlUbQeg5xgfiB16muV+0zI6lhjIxl+/6VOLolwroq5HQnIqHAakdauk+GNXlUQSTWUzEnyw+3cSeB82V+gX1/Kne+BtTtwWs7iO6UAfKT5bE57A5Hv1Fc0nmqN1qAy/8APuzYH/ACen0PFbmkeLLuzQRxyGaOPh4J8hk6DHqvp3H1rP2VOWklY2jWnHZmPd21/p0oS7t5YWJIXeuA2OuD3+oqIXbg85P4V6dp/iPS9XT7O7LHJINrQzYw2cDAPRs5xjqfSor7wZot6xcW7W0hIJa3bb0GMY5UfgKzng3vF3No4pfaR52l+c8Z+mKl/tGQ9M5Fauo+AdRtwXs54blQASp/dsTnsCSMY5ySO9c5qOnajpbn7XazW6AhQ+3KEkZADDgnHofWuWVKUd0dEasZbMuT3ssw5OPcDmquzzCcsFX681SFzJjPmEmpFuyDjLE9elTqVoXkjCdJPbg1G0vzfLnI5J5/xqA3RwMgEH2p63SEYO3PpS1HdF+K6IXLbd2O+KkWaORju6kdm/8Ar1nC4Qn7gP0OKcJo9mNjcdOeKA0NES2y4zGMe7ZH5U5713/dqpK/his9ZIiudrHj1p3nw7SNrBsUDsXPtJyFBVR3IFQkwvKQY5SpH51CJkWQOqjgfxAf1oZw/K4GeuBn/wDVQKxZjuYIhiHKZ6gjP86he+CE5LD1JFRrCrruMh6dvl5/Ogwpt5cgHrk5z+VPQVjGOxQXK7gy8lT0p9vZyTK5to5JlX7xCk7fTOB7H8q3YLeVMG3gt4MNvVlTcyn2Y5NTPZzTHdPPLJnruY4rtWFfVnE8WuiMMaXKpXzXhi3DdlpBx9QMn9KlSzt/k3yyS8/OkceBjPZifT2rYTT41/5Z5p8nlwRln8uKMdWY4A/E1qsPTW+pk8TUe2hkpYxfOFtWIbGPNlJ2/QLirqwTB2dHWEsMEwRrHx6cD2psmrWcO7y1aVgDyo4B+p7e4zVCbW7mRGNuI4VA5LHc2fb/APVWsYRWyMnOUt2aa2UYbc4Z29W5oe9tIFI81SQPupzn+lc5LcTybZbmddo58yU4UfTt+VQG7+0OUsreS7cL1jUhQPUk9Pyockt2JI331lxlYIMEH7xyRj6VnXmpEbTdXpLcDysYLd+VH88VDFbXFx81xfIgHBjtf5Fz0/Ue1TRQwafMixqkQYcOn3m/4Gck/hj6UuZv4R27kKyXT/dg8heu6cfMf+ADn8TxUpgURF3Wa4brmUAIvuEHH55oZSJ5PKKDnkovJ9eOualTIRVJIcHJVuB780ct9wv2GrKUIcxyMMcBeAM9MHsPaljnxwQwIww7/rTpvKVhKsvmFgcL0U44GM9eaqG4YR7vI2hiSWI4H4f/AF6uwiQyyGfAcspHO7nn+lKHY/6tTuJ5CgkH6nNNhMjb1jA3DnLA8fgKmiVIhvQ5Yk/d6Z9v/r0AMNt5EwYzDbwdqnpTnuFlwhbIzwO2akkuI5Sis2dxw2ACSPwquI7SOJz5xI5wSec/SgCWOd0wQinqoyoYnP8AKps+agLMS2OQOQOoqp9pg8lIkVmf2zz61LHIZC6RAfIDtKDPT2oARJjPOY5PLTnHTODn1qQeahLEvjJ+UHOR69OKr/ZpULSEfN3wMZPtUomOBE8TFSM89/8ACgRNu2QmXdk4+7nnj0I71HceTcOrOksbgAJOv34/b/aX2P4Gn+UrgKpJGCAy84pfn3IrSZzkdOORScU9xp2Izey2kkS36gI3Ed1GvyP9eMqfb9BXRad4o1DTlSNLjfB1RJTvUjGAAeuB7GsJWRUeN0Msbj5o3xtY+47fUVU+x3FqGk0wNJEOZLST7ye6+o9xz6is7uG/9f19xW56rpfjTTb8+VcMLOfP3ZW+Q9ejfQd8dcDNb7LnlSVPqK8SttRgv1PkEiZR80TN8wx1x/e/D8hW3pHiC/0xSIZC9vnAjI3oOvQDkcntj8acveBOx1mo+FNLueW0+NG2hRJbfuivOc4Hyk/UVy974I8ssbG+VjliIrgeW2OwDcqT7naP6dXY+LLC9/dzbrWXpiX7h69G+g74696vzxpIcOgP1rmlDujojUa2Z5ReaLqFgx+2W00KggByu5CSM4DDIP51QaNw+UKn8+K9cFu0LFoLiSE+meP8Kwr7S9OuARJpqFsBRLZfuyADnoPlP1I/pWbpdmaqv3R58QwOCGYg9hx/KnJM6NhmxnoK6O58I78tYX6E/MfJuR5beyhuQT2ydo/pk6hpd/pvzXdrJECQokPKE4zgMAQeM9/WspQa3Nozi9mVxK4U5I4/2ab5r7cDIPXqaiZ2EYIIB65HWkjlD5BOfxxU2KLAlcYDOw9MHNOa4CHJ3c981Aro6srFQO3Tmn+XuOAqkdc5x/KjQCVL1VBw7Z7g/wCFNF/n5g7Y/T9KQISf4QPQnJprRZHyyH8aNA1OmmW0hQv5otupyGxn8O/5VRbW1iLCM/alU8sq7SeM/wCeK50WyIwMixo0mNq85+tPZ7S1tnFzsibqBKASfw6/nXruy3PI1L9zrt7LvEe2JADwDzg+/wDhVaRizbmzNIfvAgscY65+lZx1kyuINPtDMzDjKkA/gOT+JNJNYTPhNXv1gQ/MLaBcse2dowAfrzUynb+v0Gok73tlDv8A3iE5ztj+c+2ccfrSbr66CtFBHZwvyJZzjd9Aev4A0+1NtbQbrGAQOOkk372TgepAC/gMj1NPcqXV7hnkLHcPMm5c+/c8/lzS9576f1/Xcegken2QmDXU0l7KBkmYkJ7YHU/mPpVpyXt2jMqRxE8KAEUfhx/jU4u7eK2VmkfCEgLjhMd89D+dV7swOPtCbXXbnIOSD6AE4z/nmqUFuK7K2I2basisi8gFchqkjaB5CpOZF4c5LHb357fQcUhug0SF4mUZwF37s/U8Yp5nhWRUIUynAEcb7vp0GKoRbSUGdTG0UCDmMsSGxjHY1XuyxdSHxg5JUksx9cDpTNQWZC4njQOp2kZ+6fUn1qnFMEJy4YqMhhuwT680AaLA8I6bfNy5Qc4z65pztbecChw3C4QcAflmoxfI8aOfLkI6EqQFbv8AyFQwrcWsp8xmG5txz82AeuaAFkmlt7hkSd2h/i3Hdnn0P4VMYhdRxeWACBtPIPHU464pIoZJLZpTIm1eWXJJP/1uRRvkSOKeQqrKoKxqu5T+Z/GgRNJFGImKptdyNvzDJ/r7YqnduokRWi2OcbgV2hh3x/8AXqCaee8iyv7r5wWAwN5z2wO2T+dWZ7MTGKKQzrIVwQp2kgf574poBskpitlzFJt5UrH97j/9fpRBcfvh5UkoQqREkpzg+nHfrSTYmQbViRVwCM5y2OcYweePXpTlhhZJCrh3wCAeoPH3T070AWSv2NZ5FaN/4fLf7xzg546H17H9KjlkzZIyIjvuKq8a854x3/lUm55ZsSxK06DlmyQTg+nFRRslsskSzysHyVBA4f09fSgC2lyscLLHKC5wDjHH154qFbia3LSndKsfRm5AH1Pbr/nFNljvXViQiRDLCRsdMA9Bzjp+QqdCqsvnGNyOd0R5IPooOe/v9KAG+fJPbvLI6GNgW2qwyuPxpTIwjVo5EUH5o5GIBP09DVUOocPII9+/hinzBc8Ef4VOkcRcYcBcnnOB09O3rRYCvJYW2qEu+y0vxys6E4Y9t2B3/vD8R3qJL65srpbfV4nSVRlZ15LD1ODhvqOfc1eaJioa3kTHAKq2crjB6d6RmiMTWV0v2qBzuUFsbD32Meh/Q+lZODWq/r0/y2KT7llB9pj3qRIrH5HRsqf8+/NWLPVtU05QsNyyovCxscpjOTxzj8Md65yW1vdJf7XpcnnWhOJVxyv++vb61cs9Rtr5gqOYJx/yyY8E/wCyf6H86FJW12/rcDvNN8TWtwFiv4jDc/3n+dCeOh7dfw9a3PtMbxBo23AjIx0I9q80yjTurcSxj5l6EdB/kVNaardWWdmVRWwY8ZBz16/T64odNPVDU+53c6LIM7Mj2qujS25zBM8f0OKzIPEVsGSK5LQSHADEfKTx+Wff061qmUyJkbWU85HcVm4tFp3Mu50/TrrifT41bAUSWuI2Azn7v3T9SD/KsS98KJIxfT76Mn5mENwuxvYA8gntztH9OnIiY5b8xTGtwwyrB/Zqh04s0jUktmcDd6RqVh/x8WsseCBuHMZJGRhh8pPtnsaqCR4pVyzAdRv6Hj1r0hWnt2BhkeM9sHIqjeWNlfKPtdmAyqFWS3YoVAOeg+XPbkH+VZOj2NY111RxgLywqUBUNyvz5ogaRiylhuXgBuM+w9a3Lvw0JQx0++jcYbMVwu1sdgGHBJ6ZIA/pmXOlXmnKFurMBdwXzWPy5IzgMDjP49qylBx3NozjLZmIh1jUo2ltkW1txw0w+UD6uT+mfwog0vToVLvJNfz4zhBsjz7nqfwH41o3bG6kSa6bcAAoDvuwB7dAPoO1RLIJo2SEEAMUUtlQw9cf0r0VCT309P8AP/hjzbon8xwi20DJaxsMyLAmwj2z94/jVdrK3spfmt5dz8gycNj1ORVgXcForCCPEp+/85OT7n/I96z5b+9M7OF8yRwT1O1B35J5/GtIxS2JbLu9pIzMrgOvAGwde2OP1/8ArVBdSyXUEbnfkLg7VGCemRxz6e1OsZ0WNJJJ4iTkEISNp+vr+dTO4lgLOibfv5XjA9M9fxpgU43WJRCYGYHlv4iT34/DrxVxpvssq4n3kqMxMuGA/wAcflWfbC53yGKOOOMceY/OfXAzz6d60IbWzjVC80ayuvz8gMx6kenp60ARTangFVVPK2jeqL3z3JPNV4nkSQTmAiaU/I27Ix049amuEe5u/Itw6RnouAfxAHal2iPy4QiKSdihUx/wJuxPPrQBSuzPcSGASbUGFVwAP5jP8qnh0+K3If5bpUYbzI+R+OOlLLp6QzYiuPMffgjB2IPwH8q0Le9ltklXZ5rHbwvK+nTj169fzoAkur2xSZYwEKbN+IgDhsdiDxVZpvNuWeHezFQQJGzyOpPpxUL2jWVyRIRLKAMoXHyfWpI7p7i5iS1SS2UdArAhm/Qnv/nigBHhE1xiZWxHyqoQFyfU+gqWBXtl8+LEgbB3sfvAEduw60XsdsbNU85mdGy442k55/zz9ahLR26t58hLsCyqvSgC1dWhknjKTN9pc4SNeAR3OPrn9apxxrbxzoqGSUPgsRnIz6/UDiomkZJ5GcyJNtyuT0B+vNPWJVgjkSCVx947WAyAewxQIdcXlq0zEWallC7io2oMdV4/nVCfU18lfLwgLFG3DOwH0HbFXruKWWMSQHzY2b5lC4/lUFusDRyedAixOOFYZKn1FMCWxkBPnpl8HBAGMHHBH61cRBNJ5kxKMuMOGDbep5I/GqxuGsbF47Vl8tTlCYsNIDx1qXSNOa5tnuXZY7dVG5Rxkk9Gzx68+9AXI5ArqyqxVs5BZvlbv19qls5vPZlKgseFdFDEcetLezRJICjKUYDKEEA9s49f51WjgdbrYCkYQ7mGcAY9P8KAL0gMcHlsNpIBLqfmz2/z71OlxFAqld/nMNrOc5zjKmppZJY44J5LdWhLFtzL8rduT0GP8Kjub3T0TzJbR0ViFLIuVTr1559elICvDcKmWuogn8W5PX16/UUCeCbbJtZ25wNx4I9KbLKi24ESRcuG24J3I3IyOvbtVMFUYws7wOHwYypGD/LtTEaIS6trkTWp2Fu64YOCMkHnB78H9Kp3Gn2uvgS6WVg1Ifet1bashGPu56H/AGT+B7VZCtOhVZR56fdYkgyYyQDxz3/EiqFxZNfSYlhkiuySMFQgbrySMc1EoXfNHR/1v3Gn0ZFbatdWL/ZdTidxGdpJ+WSP2yeR9DkfStuBkuIWuLaZZcHJKkblHuhOR+GR7mq5ltr6ySDXfN/drtS76yLjgZOfnX9RWPc6fd6DMt3bzEQOf3VzE2VcfX+YrOLs7bP8H6FM6AogcSmIuv8A0y5IB9sj8aS3uXsrrfbTPGW7HjceccdMc9OevHrWfYatY3p23Ci0uSuA6cRsfUgfd+o49hVtYpo5gsse1uWU4yGA7gjjHvWild2e5Ox01rriuP8ASAoz/wAtIwSMYzyOv8/wrUjkjmjDxujKejocg1xJiaM+aiOu8ctGOv4VHZ6iLS5LxzSIxGXRhgNj+dJ077FKXc7x2cDDDOPSqzsSev4GqVprAuVUNtGRkYJ/yMVdWRJRlHVwOCQc4NZuLRaaYx41fGRz7URyXFuwMU7LjoDTycA7TTd7Hjj8aQ7nAWz6hqDtbQBpWdi0jBGLYxx1wAP8noKVLPlBKuGDEbYjvbHc8H+fX6VejEdgGi1DUJWgYFZI1cgsBzgjGAc8YO71wKyJtS+0Xb+UjLbBt4jn+fn/AHc/r+g6DoMhNR1C0t41iWMZzu3AkFh/unmmQ2Umo4cQzRRs2MD5iCfbGf0p8EVjAwcpGJHIPmNtkx6cdh7VpfJb2IM5aCF1JLBcbs/3FBGe3UnrzQBUsNGla5kij2iGNsSS7kBA79T8vSrn7zEcYA+QkxiY/Lt9eMZP1FZ0ASaJY0eOMudxG4jcT0z0A7cYzzVu2sjcM8krPI6jAVOdpzgd/wD9X6Urg0LJJugKWyTSkAGR+w9s9qzxYXMsvmOjhgcru44/GpreXdK7TK/kR8F40IVj0A7Vs28tvDBK5UE52RCSYDaCPvc+mPSmBFbWcVnBLe3+5CBtIOeTjgZ4zmq9vf2yLtlBDBcrvXJJPcD/ABNaS3Z1YZtgkjQAMzSqfLQ4xyxwP58+tZtxMFu5bhpftDsuMqChPHY/3fwoAdG6vOHwqsqDakbckE8biBgH9alm2xQGWOPrFlXldQ3J6hR1Poe2KxrWO4W6kKMoVcMULZBP9TWq07C6acSBZCMBO/0HX86AM4tCkflrOBKxyzuOmByPTr7fjVjTka6gkmnlVrRTtkDDpz64yM47elVUs0VN86XH2gSALjspPYY5JPFSxytds8qQKlopBZ3AHXgZJ7nk0wIbye2vdQht7JGgtjtLAkqmc4z14HXn/wDXV+6iEFzObV4HjCjE2SQxwCVGc5IyM/Smw6QZ1JQLbwsclCCWHp6VpW+mx28hcSuz7dq4UYVcDpnPPXmgV0YX2kGa4Z1QnDKWdgS57fkO465rSj2JEtvFKsyeV8zoDhSRjHr1x3q2NJsgCPJySOSWIPbnOevHap4ra3gjVIoIlVVCgbQeB9aLCuYFkLiCXdGqqoGCrAkZ7t161aDiVDFHFGI4iX3SrguP4R+NbIYKpURxgdceWuP5UwpEVKm3g2k5I8pcE/lRZhdGAJhKjQZMapjdtHAJ9fT0pZLx7SyeHzQVVhhF5Qn1NbjQW5620H/fpf8ACkNvavEYntYmU9Rgj+RHpQF0cuyypCHUuFY4wSABgfX6fWr1ut0LL7UkTfuztlyuUccDr07jv3FazWNo1v8AZxE0cYOQsbfzyDmkGnwGMotxIC4VWZ1zjBznI78AdKAuD3m+yCwsURgWVPLUlGbgqCT93ByKzvtjWV5JFOvm+YgKruJVDt442nPWrD6ddGRwHieN+WIYZz2wDg1WmjuI7yKeWJZMKD/dz05470ICy8CwMWtrlmgKZjypDDjnIzjPfPqKmnYtbRC4O1pUUkug5UAjqT7Dtnk4qQtG+nn7WCgUnaX5IcjrkdOAKrI0zK1tFBvlVFKMSWOAdx49D70APu3XEXkyjMZGGwFJAzjk857fhT2lNzDGu4x4O0gR5ZW7859vpVeO6aQSD7PHHJvBUkkmI9CAPSrLNNL5ktvbRJOApbyxnCgYPXnPGaAK8Al+0pBPHJ8x2rIWUK2eD14NTW9xcafDNAbYXEGQrW5fdG/r1Jwfpimfara/gitVgCSAZfcxUFvUH1/wqGSYWSGG4t/9JJByc/OO3TvUzgpq0kCdtivP4cgvbaS/0gybRkvZuR5sf0/vD9azNP1e8sCbeVTPbAnMcg4B9R/dPuK3Wllh1BZw00Dkhg8XOMD6fh/nNTXUdl4hkZbh4oL9QD58alBJ/vADg89ayacVaesfxX9d9y9HtuJaul8caezSpjcbaTHmL/ukfe/n7U2UWt2SwVQ2BywIIOMHoa53UtNvNGnEdyDDMh3Alxhu4I9fqK1dO8R+fuF/lH6/aY1Bb6sOj/Xg+/arTaV1qv6+/wDrcTRpK0lmgVB5iIO3T8MH9aaLm489ri2uCrAAFCwAP6UoMikXEzxTwyDas0Q/due2emD9RSXkcasrwCMEAbgWxjnp7H2JI6VaaktBbGtZaq2xftgGScb14/TPP1H5VqxyxyJvjkDAHBxzj2Pp1rmo4JII3/1jW8mH5O4Zwe4Pv+tQfbZLWUSW0jbSPuDPQeoPGPwqXBPYakZLaXZiNGa58yQjLu7BQR7FgD+mPrUSNDNIbSNIxGASZl6t+LYyPoOam1CGTUJo/IiYEZ3liS5GepJ7/wAqahhjLeWvmbSMukRIU+hz/n61QE7RxvGjidwAfLCHO1eOrYyWP4fjSxaVLbSR3F1AXM/MQnwoceoBz6ehHaiIyTyFbi5mS0zgkEYz2GDgAZ7Dp1qFoLozkwO+5ZMNM0mFDH1fGPxNAFu2060lu7h5Vhi2/wASAkj6BsBOeMtipLzWrWGT7Pas7FlGx1jwDjg4PUDPXH3sdupltoJryzkS4i85I8CKZ5XYRjP8KgEnPrjjmnnT9RmlupT5QncbY5pXIZAD1UJkDgYwfX8gRnLevMoto7dnkXIVpBywHIHXA5znH5ZqJpbg3CXckUQkQfLCi8ntlh/CucdeegrrRZWUDQzDfvtxlJiQjKeucjnI9c1C9vY/cFsCu7BAGAOM9+o+metNITkcxFe3i2q2jOZC5LIA4HlsTgnAJP6ZP0xmxY6dqPmmZYGKFNuD8oOf97HH4GuihaKFF8uCONyPm2f44GfrTzK57GnohXbMe28PSxSxs8sSqvLqSXDnPTopA6Dr+NWo9FgXa8tzMXH9wgD6DIJqw8zj0qtLcbfvScnOAO/+NK4a9S00VkiBPIVwqhVEzGQKAScKGJxySePWoHhjmuzcSs8z8bA5+VP90dvrVP7T02ITnoW+Ufrz+lSK0rY3OEHoo5/M/wCFA7GgoAIOVX8KeoiOTuLY645x+VVII/MJKl2IHJLBBVaz1G3uri5t4IQstq+2UTSdzyCMnBHX/wDVip1Y9DVTyC2AhJ9+P508iMceSD/3z/jWfLdup2+VCrDuFB/nmmDUJ16Mg/7Zr/hS5ZDvEvma3EhQxIHABKkrkA5wevsfypGuLVFLPHGqqMklkAA/Oqf9qXQ/5aD/AL4X/Cg6pdn/AJa/+Oj/AAp8shc0TTCof+XYf+O/40jxRAAtB1/u4P8AKsw6lcnq6n6xr/hQuoSBsvHBIPRoh/TFLlkPmiaQt7dxxFIP+AtUYtrZx8k/P+8DVU6q5GPJiAHQKWGPyarC6j5kPy29wIMfMwcOMd+q8/nR76D3WK2nP/BIrfXiq81lKB88O5c+gYVcjuLfYsmGjjABBaJk3Z91JH6VJFKWACSeZ3+VlfC/Thh+VLnYcqMWS1hlcsykMRgkH2x0qaNFghaO1xASu0umdx6Z5JPXn8/wrUd45CFmiUsR26/kcN+lRGyhkJ8qQqQcFeuP61XOnuLla2Mn7Lbq8WUkaRvmkfGe+CM/THf19M06EI+mm5jeHY7mMeXOFbIbIzjqOvJ449RV17O4j6KHH+yaquimVGkjHmR52ll5XPBx6Vas9mS21uUNSQ2BRLmAFm3rIhzkduCe4P45yCKdAY70R/ZLkSQuiGVTksNpwVI56dc9P5VI2n2zII/3oXdvYl95b/Z+bPB9qIbAQPFMqwSTgkswUxKoC4XCgnc3fJI5xnqTTsF0LHHLG0vlSEKSwL+aCVce/ORx3HP4E1WezuZ9ouWXzVYRplQS2emXH9RxjtT7pClzaXR/clgWlXAyCGIzkZUZx0ZiTn3xSyzK8kTwP5jTDDQ+TncQRhdvIzz6DpxjIpahoWjqMiwJo2p4nhClSdgVoj0BB/Q5x78dcDVvCr20Rv7W4+2WBGRNCuTFn++vVfr0+nSugjee88jzJLWPzBgFo8Eeg54OR6nBx7VQ0y9fTb2RoZzDuyBhGeGTvhl6j/8AX61zypyj71LR9uj/AMn5/gy1K+kjmtP1i/0skwt8uMOjqGDL6EYII+tddpusw3OyXT5Yre6ZNhtZiPLfnO1WP8m/A9qS70fTvEStc6KyWuqAbmsX4jlHfy/y6fyrjp7b7HK8MqSxXC/fjkOOfalCUazdvdmt1/n3Xn9zG015o7X7VM9w4eKW2mVsSQkbRwOeo/T8qzr2AzTq8FwWC4+QKcofU9/xqppuutJbpaajG9zBgrG3mbZI1P8AdbuPY5FaQWQQebbSm8sgw5VRuj9mXt9RWqm07T0/Im3Y/9k="

def img_preprocess(img):
    img = img[60:135,:,:]
    img = cv2.cvtColor(img, cv2.COLOR_RGB2YUV)
    img = cv2.GaussianBlur(img,  (3, 3), 0)
    img = cv2.resize(img, (200, 66))
    #img = img/255
    return img

image = Image.open(BytesIO(base64.b64decode(b64)))
image = img_preprocess(np.asarray(image))

print(image.shape)
plt.imshow(image)

cv2.imwrite("img/demo.png", image)

model = load_model('model.h5')

# convert tensor
img_ = np.array([image])
print(img_.shape)

float(model.predict(img_))
